+++ src/modules/Opportunities/components/opportunity-registrations-table/script.js
@@ -12,23 +12,54 @@ app.component('opportunity-registrations-table', {
         return { text }
     },
     data() {
-        return { }
+        return {
+            filters: {},
+            resultStatus:[],
+            query: {
+                opportunity: `EQ(${this.phase.id})`,
+            },
+            selectedCategory:null,
+            selectedStatus:null,
+            selectedStatus:null,
+            selectedAvaliation:null
+        }
     },
 
     computed: {
-        headers () {
-            return [
-                { text: "Inscrição", value: "number" },
-                { text: "Agente", value: "owner.name", slug: "agent"},
-                { text: "Categoria", value: "category" },
-                { text: "Status", value: "status"},
-            ];
+        statusDict () {
+            return $MAPAS.config.opportunityRegistrationTable.registrationStatusDict;
         },
-        query() {
-            return {'opportunity': `EQ(${this.phase.id})`}
+        statusEvaluation () {
+            return $MAPAS.config.opportunityRegistrationTable.evaluationStatusDict;
+        },
+        statusEvaluationResult () {
+            let evaluationType = this.phase.evaluationMethodConfiguration ? this.phase.evaluationMethodConfiguration.type : null;
+            return evaluationType ? this.statusEvaluation[evaluationType] : null;
+        },
+        statusCategory (){
+            return this.phase.registrationCategories;
+        },
+        headers () {
+            let itens = [
+                { text: "Inscrição", value: "number" },
+                { text: "Categoria", value: "category" },
+                { text: "Agente", value: "owner.name", slug: "agent"},
+                { text: "Status", value: "status"},
+                { text: "", value: "options"},
+            ];
+
+            if(this.phase.evaluationMethodConfiguration){
+                itens.splice(3,0,{ text: "Resultado final da avaliação", value: "consolidatedResult"});
+            }
+
+            if(this.statusCategory.length == 0){
+                let categoryInxed = itens.findIndex((item) => item.text === "Categoria");
+                itens.splice(categoryInxed,1);
+            }
+            return itens;
         },
         select() {
-            return "id,number,category,status,createTimestamp,statusesNames,consolidatedResult,owner.{id,name,files.avatar}";
+            return "number,category,consolidatedResult,status,singleUrl,owner.{name}";
         },
         previousPhase() {
             const phases = $MAPAS.opportunityPhases;
@@ -38,7 +69,52 @@ app.component('opportunity-registrations-table', {
     },
 
     methods: {
-        statusToString (status) {
+        setStatus(selected, entity) {
+            entity.status = selected.value;
+            entity.save();
+        },
+
+        clearFilters(entities) {
+            this.selectedCategory = null,
+            this.selectedStatus = null,
+            this.selectedStatus = null,
+            this.selectedAvaliation = null
+            delete this.query['category'];
+            delete this.query['status'];
+            delete this.query['consolidatedResult'];
+            entities.refresh();
+        },
+
+        filterByCategory(option,entities) {
+            this.selectedCategory = option.value;
+            this.query['category'] = `EQ(${this.selectedCategory})`;
+            entities.refresh();
+        },
+        filterByStatus(option,entities) {
+            this.selectedStatus = option.value;
+            this.query['status'] = `EQ(${this.selectedStatus})`;
+            entities.refresh();
+        },
+        filterAvaliation(option,entities){
+            this.selectedAvaliation = option.value;
+            this.query['consolidatedResult'] = `EQ(${this.selectedAvaliation})`;
+            entities.refresh();
+            
+        },
+
+        consolidatedResultToString(entity) {
+            if(this.phase.evaluationMethodConfiguration){
+                let type = this.phase.evaluationMethodConfiguration.type;
+                if(type == "technical"){
+                    return entity.consolidatedResult;
+                }else{
+                    return this.statusEvaluation[type][entity.consolidatedResult];
+                }
+            }
+            return "";
+        },
+        
+        statusToString(status) {
             return this.text(status)
         },
         isFuture() {