+++ CHANGELOG.md
@@ -54,7 +54,11 @@ e este projeto adere ao [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 - Implementa a criação de usuário atravéz do fake-autenticate
 
 ### Correções
+<<<<<<< HEAD
 
+=======
+- Corrige erro da instalação do npm após o node. 
+>>>>>>> feature/technical-result-apply
 - Corrige erro na listagem de inscrições de um edital quando não exitia nehuma inscrição
 - Corrige erro que ocorria no momento da publicação de resultado final ocasionado pela tentativa de enviar os e-mails
 - Corrige validação de metadados preenchidos com valores que o php, com cast para boolean, resolve como false (Ex: 0, [], '')
@@ -91,9 +95,13 @@ e este projeto adere ao [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 - Ajusta botão de sair do modo de edição das entidades para que o mesmo não salve a entidade
 
 ## [7.2.10] - 2023-12-06
+<<<<<<< HEAD
 
 ### Correções
 
+=======
+### Correções
+>>>>>>> feature/technical-result-apply
 - Redireciona o usuario para a primeira fase quando o mesmo tenta acessar a oportunidade com o id de uma fase posterior
 
 ## [7.2.11] - 2023-12-07